refactor: give moganet its own CS/CF objectives
